Correctly deal with TimestampNote
