+++ src/main/java/hudson/plugins/timestamper/accessor/FreestyleTimestampLogFileLineAccessor.java
@@ -2,16 +2,11 @@ package hudson.plugins.timestamper.accessor;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.google.common.io.CountingInputStream;
-import hudson.console.ConsoleNote;
 import hudson.model.Run;
 import hudson.plugins.timestamper.Timestamp;
-import hudson.plugins.timestamper.TimestampNote;
 import hudson.plugins.timestamper.io.LogFileReader;
 import hudson.plugins.timestamper.io.TimestampsReader;
 import hudson.plugins.timestamper.io.TimestampsWriter;
-import java.io.ByteArrayInputStream;
-import java.io.DataInputStream;
 import java.io.IOException;
 import java.util.Optional;
 import org.kohsuke.accmod.Restricted;
@@ -20,8 +15,7 @@ import org.kohsuke.accmod.restrictions.NoExternalUse;
 /**
  * Implementation of {@link TimestampLogFileLineAccessor} for Freestyle jobs. In contrast to
  * Pipeline jobs, Freestyle jobs do not embed the timestamps in the main log file. Rather, they use
- * a separate timestamps file in {@link TimestampsWriter}, optionally adding a {@link ConsoleNote}
- * based on the settings in {@link TimestampNote}.
+ * a separate timestamps file in {@link TimestampsWriter}.
  */
 @Restricted(NoExternalUse.class)
 public class FreestyleTimestampLogFileLineAccessor implements TimestampLogFileLineAccessor {
@@ -52,50 +46,7 @@ public class FreestyleTimestampLogFileLineAccessor implements TimestampLogFileLi
     public TimestampLogFileLine readLine() throws IOException {
         Optional<Timestamp> timestamp = timestampsReader.read();
         Optional<String> logFileLine = logFileReader.nextLine();
-
-        // This implementation relies primarily on the timestamps embedded in the timstamps file,
-        // falling back to looking for ConsoleNotes only if necessary.
-        if (logFileLine.isPresent() && !timestamp.isPresent()) {
-            timestamp = readTimestamp(logFileLine.get());
-        }
-
         return new TimestampLogFileLine(timestamp, logFileLine);
-    }
-
-    /**
-     * Read the time-stamp from the {@link ConsoleNote} in this line, if present.
-     *
-     * @return the time-stamp
-     */
-    private Optional<Timestamp> readTimestamp(String line) {
-        byte[] bytes = line.getBytes(build.getCharset());
-        int length = bytes.length;
-
-        int index = 0;
-        while (true) {
-            index = ConsoleNote.findPreamble(bytes, index, length - index);
-            if (index == -1) {
-                return Optional.empty();
-            }
-            CountingInputStream inputStream =
-                    new CountingInputStream(new ByteArrayInputStream(bytes, index, length - index));
-
-            try {
-                ConsoleNote<?> consoleNote = ConsoleNote.readFrom(new DataInputStream(inputStream));
-                if (consoleNote instanceof TimestampNote) {
-                    TimestampNote timestampNote = (TimestampNote) consoleNote;
-                    Timestamp timestamp = timestampNote.getTimestamp(build);
-                    return Optional.of(timestamp);
-                }
-            } catch (IOException e) {
-                // Error reading console note, e.g. end of stream. Ignore.
-            } catch (ClassNotFoundException e) {
-                // Unknown console note. Ignore.
-            }
-
-            // Advance at least one character to avoid an infinite loop.
-            index += Math.max(inputStream.getCount(), 1);
-        }
     }
 
     @Override
